EditモードにおけるJointのlockRadiusの編集を可能に

+++ src/Edit.java
@@ -49,31 +49,51 @@ public class Edit extends GameState
     public void update(GameContainer gc, int delta)
             throws SlickException
     {
+        objectPool.collisionDetection(gc);
+        grid.update(gc, objectPool.camera.getX(), objectPool.camera.getY());
+        mousePointer.setPointer(grid.getGridCenterAbX(gc.getInput().getMouseX()), grid.getGridCenterAbY(gc.getInput().getMouseY()));
+        mousePointer.update(gc, objectPool.camera.getX(), objectPool.camera.getY());
+
         if (gc.getInput().isMouseButtonDown(Input.MOUSE_LEFT_BUTTON))
         {
-            switch (mousePointer.type)
+            if (mousePointer.isCanPutObject())
             {
-                case GROUND:
-                    stageData.addGround((int) mousePointer.abX, (int) mousePointer.abY, mousePointer.ground.getType());
-                    break;
-                case JOINT:
-                    stageData.addJoint((int) mousePointer.abX, (int) mousePointer.abY, mousePointer.joint.getType());
-                    break;
-                case BACK_OBJECT:
-                    stageData.addBackObject((int) mousePointer.abX, (int) mousePointer.abY, mousePointer.backObject.getType()
-                            , mousePointer.backObject.getLayer());
-                    break;
-                case CHERRY:
-                    stageData.addCherry((int) mousePointer.abX, (int) mousePointer.abY);
-                    break;
-                case HEART:
-                    stageData.addHeart((int) mousePointer.abX, (int) mousePointer.abY);
-                    break;
-                case DELETE:
-                    stageData.deleteObject((int) mousePointer.abX, (int) mousePointer.abY);
-                    break;
+                switch (mousePointer.type)
+                {
+                    case GROUND:
+                        stageData.addGround((int) mousePointer.abX, (int) mousePointer.abY, mousePointer.ground.getType());
+                        break;
+                    case JOINT:
+                        stageData.addJoint((int) mousePointer.abX, (int) mousePointer.abY, mousePointer.joint.getType(),
+                                mousePointer.joint.getLockRadius());
+                        break;
+                    case BACK_OBJECT:
+                        stageData.addBackObject((int) mousePointer.abX, (int) mousePointer.abY, mousePointer.backObject.getType()
+                                , mousePointer.backObject.getLayer());
+                        break;
+                    case CHERRY:
+                        stageData.addCherry((int) mousePointer.abX, (int) mousePointer.abY);
+                        break;
+                    case HEART:
+                        stageData.addHeart((int) mousePointer.abX, (int) mousePointer.abY);
+                        break;
+                    case DELETE:
+                        stageData.deleteObject((int) mousePointer.abX, (int) mousePointer.abY);
+                        break;
+                }
+                objectPool.init();
             }
-            objectPool.init();
+
+            mousePointer.canPutObject = true;
+            if (objectPool.wire.jointLockedNum != -1)
+            {
+                mousePointer.canPutObject = false;
+                int jointLockRadius = ((int) (objectPool.joints[objectPool.wire.jointLockedNum].getDiX() - gc.getInput().getMouseX()))
+                        / Ground.WIDTH * Ground.WIDTH;
+                Joint joint = objectPool.joints[objectPool.wire.jointLockedNum];
+                joint.activate((int) joint.abX, (int) joint.abY, joint.getType(), jointLockRadius,
+                        false, 0);
+            }
         }
 
         if (gc.getInput().isKeyDown(Input.KEY_LCONTROL) || gc.getInput().isKeyDown(Input.KEY_RCONTROL))
@@ -101,15 +121,13 @@ public class Edit extends GameState
 
         objectPool.moveGrounds(stageData.getGroundXs(), stageData.getGroundYs(),
                 stageData.getGroundTypes(), stageData.getGroundPositions());
-        objectPool.moveJoints(stageData.getJointXs(), stageData.getJointYs(), stageData.getJointTypes());
+        objectPool.moveJoints(stageData.getJointXs(), stageData.getJointYs(), stageData.getJointTypes(),
+                stageData.getJointLockRadiuses());
         objectPool.moveCherries(stageData.getCherryXs(), stageData.getCherryYs());
         objectPool.moveHearts(stageData.getHeartXs(), stageData.getHeartYs());
         objectPool.moveBackObjects(stageData.getBackObjectXs(), stageData.getBackObjectYs(), stageData.getBackObjectTypes()
                 , stageData.getBackObjectLayers());
         objectPool.update(gc);
-        grid.update(gc, objectPool.camera.getX(), objectPool.camera.getY());
-        mousePointer.setPointer(grid.getGridCenterAbX(gc.getInput().getMouseX()), grid.getGridCenterAbY(gc.getInput().getMouseY()));
-        mousePointer.update(gc, objectPool.camera.getX(), objectPool.camera.getY());
         counter++;
     }
